fix: remove clear_layer_data call
The `clear_layer_data` call was accidentally introduced with the Synthetic
PoRep. The call does under the hood exactly what `clear_cache` is doing.
It was already removed from the FFI via [PR 436], so that we can remove
it here as well. This reduces the API surface, which is generally a good
idea.

[PR 436]: https://github.com/filecoin-project/filecoin-ffi/pull/436

BREAKING CHANGE: `clear_layer_data` was removed, use `clear_cache` instead.